refactor/breakup router
